splitting out ts code into its own file to ensure maximum type safety
